+++ _ide_helper.php
@@ -3,7 +3,7 @@
 
 /**
  * A helper file for Laravel 5, to provide autocomplete information to your IDE
- * Generated for Laravel 5.8.33 on 2019-12-01 01:00:18.
+ * Generated for Laravel 5.8.33 on 2019-12-09 15:50:40.
  *
  * This file should not be included in your code, only analyzed by your IDE!
  *