+++ database/migrations/2019_11_22_002451_create_customer_comments_table.php
@@ -23,6 +23,7 @@ class CreateCustomerCommentsTable extends Migration
             $table->integer('parent_id')->unsigned()->nullable();
             $table->foreign('parent_id')->references('id')->on('customer_comments')->onDelete('cascade');
             $table->softDeletes();
+            $table->integer('level');
             $table->timestamps();
         });
     }