+++ database/seeds/seeder_tests/TestCustomersSeeder.php
@@ -19,6 +19,8 @@ class TestCustomersSeeder extends Seeder
         DB::table('customers')->insert([
             'name'            => 'Customer A-WNY',
             'type'            => 'Individual(s)',
+            'city'            => 'New York',
+            'line_1'          => 'Line 1',
             'organization_id' => 2,
             'deleted_at'      => null,
             'created_at'      => now()
@@ -28,6 +30,8 @@ class TestCustomersSeeder extends Seeder
         DB::table('customers')->insert([
             'name'            => 'Customer B-Spring',
             'type'            => 'Business',
+            'city'            => 'New York',
+            'line_1'          => 'Line 1',
             'organization_id' => 9,
             'deleted_at'      => null,
             'created_at'      => now()
@@ -36,6 +40,8 @@ class TestCustomersSeeder extends Seeder
         DB::table('customers')->insert([
             'name'            => 'Customer C-WNY',
             'type'            => 'Individual(s)',
+            'city'            => 'New York',
+            'line_1'          => 'Line 1',
             'organization_id' => 2,
             'deleted_at'      => null,
             'created_at'      => now()
@@ -44,6 +50,8 @@ class TestCustomersSeeder extends Seeder
         DB::table('customers')->insert([
             'name'            => 'Customer D-WNY',
             'type'            => 'Individual(s)',
+            'city'            => 'New York',
+            'line_1'          => 'Line 1',
             'organization_id' => 2,
             'deleted_at'      => null,
             'created_at'      => now()