+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Customer. Field "City" is implemented. Fields "City", "Line 1" are required.
 
 ## [0.17.0] - 2019-11-21
 ### Changed